Updated with new buttons on `ViewTasks` and confirmation buttons

+++ src/components/ViewTasks.vue
@@ -35,10 +35,14 @@
 });
 
 const removeTask = (key) => {
-        alert('You are about to remove this task. Are you sure?')
-        deleteDoc(doc(db, "task", key)).then(() => {
+        if (confirm('You are about to remove this task. Are you sure?')) {
+            deleteDoc(doc(db, "task", key)).then(() => {
             delete dataList.value[key]
+        }).catch((error) => {
+            alert(error.message);
         })
+        }
+
     }
 </script>
 
@@ -52,7 +56,11 @@ const removeTask = (key) => {
         <tbody v-for="(entry, key) in dataList" :key="key">
             <td>{{entry.taskName}}</td>
             <td>{{entry.startDateTime}}</td>
-            <td><button @click="removeTask(key)">Delete?</button></td>
+            <td>
+                <button @click="removeTask(key)">Delete?</button>
+                <button>Manage Volunteers!</button>
+                <button>View task!</button>
+            </td>
         </tbody>
     </table>
 </template>
